pyproject.toml
[tool.poetry]
name = "django-mini"
version = "0.1.0"
description = ""
authors = ["meoyoug <eodyd426@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-restframework = "^0.0.1"
django-extensions = "^3.2.3"
setuptools = "^72.2.0"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"
pillow = "^10.4.0"
django-cleanup = "^8.1.0"
redis = "^5.0.8"


[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21.7"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120